refactor: dedupe starter scripts, split oversized files, shared script tsconfig
bootstrap/probe now share sibling modules (shell, vars, flags, messages,
provision) instead of carrying private copies; the wizard's prompt helpers
move to prompts.ts and build.ts's pure config transforms to transform.ts,
bringing every file back under the ~300-line rule. A parity test pins the
wizard's dev-vars parser to bootstrap's byte-for-byte, the three script
workspaces extend a shared @apollo/typescript-config/script.json, and the
generator's directory walker uses recursive readdir.
