+++ packages/typescript-config/package.json
@@ -2,6 +2,5 @@
   "name": "@apollo/typescript-config",
   "version": "0.0.0",
   "private": true,
-  "license": "UNLICENSED",
-  "files": ["base.json", "react.json"]
+  "files": ["base.json", "react.json", "script.json"]
 }
