+++ tooling/create/tsconfig.json
@@ -1,17 +1,9 @@
 {
-  "extends": "@apollo/typescript-config/base.json",
+  "extends": "@apollo/typescript-config/script.json",
   "compilerOptions": {
-    "module": "esnext",
-    "moduleResolution": "bundler",
-    "target": "es2022",
-    "types": ["bun"],
-    "lib": ["es2023"],
     "paths": {
       "@/*": ["./src/*"]
-    },
-    "noEmit": true,
-    "strict": true,
-    "skipLibCheck": true
+    }
   },
   "include": ["*.ts", "src/**/*.ts", "__tests__/*.ts"]
 }
