+++ tooling/starter/assets/scripts/flags.ts
@@ -1,0 +1,7 @@
+export function readFlagValue(
+  argumentList: readonly string[],
+  flagName: string,
+): string | undefined {
+  const flagIndex = argumentList.indexOf(flagName);
+  return flagIndex >= 0 ? argumentList[flagIndex + 1] : undefined;
+}
