+++ tooling/starter/assets/scripts/messages.ts
@@ -1,0 +1,3 @@
+import { z } from 'zod';
+
+export const serverMessageSchema = z.object({ type: z.string() }).passthrough();
